# free speedup by utilizing caches better