@YichengDWu 2024-04-18
I observed consistent speedup on three different 4090s.

 The principle is simple: when you execute a kernel and the data for its last batch of operations is in the last column, the next kernel, B, starts operating directly from the last column rather than the first column, resulting in a higher cache hit rate.