@karpathy 2024-04-18
But which kernel is this re-utilizing from? The previous kernel is a cuBLAS matmul.