@YichengDWu 2024-04-19
Yes, both before and after are cublas matmul kernels, and it's not possible to know the details inside the kernel. This speedup is more of a result obtained through experimentation. Note that after the softmax backward operation completes, the cache retains the part of the matrix close to the upper left corner, which benefits the matmul operation that immediately follows.

I call it a 'free speedup' because there is no reason for it to result in a performance loss.