@karpathy 2024-04-19
merged in 315af5f, slightly modified with comments ty. it does help about 1ms/iter on average from me, from 193ms -> 192ms. might be more on other GPUs